Move MakeEdgesCTE to csr file
